MEPTS-332 added patient  in the data-set with  encounter=21 and  obs concept_id=23944 and value_coded=1366

+++ api/src/test/java/org/openmrs/module/eptsreports/reporting/intergrated/library/cohorts/TxCurrCohortQueriesTest.java
@@ -138,7 +138,7 @@ public class TxCurrCohortQueriesTest extends DefinitionsTest {
     EvaluatedCohort evaluatedCohort = evaluateCohortDefinition(cd, parameters);
 
     assertEquals(1, evaluatedCohort.getMemberIds().size());
-    assertTrue(evaluatedCohort.getMemberIds().contains(new Integer(1005)));
+    assertTrue(evaluatedCohort.getMemberIds().contains(new Integer(12594)));
   }
 
   @Override
